refactor nft handle attempt
