+++ src/features/create-drop/contexts/CreateNftDropContext.tsx
@@ -5,30 +5,23 @@ import { zodResolver } from '@hookform/resolvers/zod';
 import * as z from 'zod';
 import BN from 'bn.js';
 import {
-  getEnv,
-  generateKeys,
-  createDrop,
   formatNearAmount,
-  parseNearAmount,
   addToBalance,
 } from 'keypom-js';
-import { get, set, update, del } from 'idb-keyval';
-import { pack } from 'ipfs-car/dist/esm/pack';
-import { MemoryBlockStore } from 'ipfs-car/dist/esm/blockstore/memory';
+import { get, set, update } from 'idb-keyval';
 
-import { urlRegex, MASTER_KEY, NFT_ATTEMPT_KEY } from '@/constants/common';
+import { urlRegex, MAX_FILE_SIZE, NFT_ATTEMPT_KEY } from '@/constants/common';
 import {
   type PaymentData,
   type PaymentItem,
   type SummaryItem,
 } from '@/features/create-drop/types/types';
 
-const WORKER_BASE_URL = 'https://keypom-nft-storage.keypom.workers.dev/';
-const MAX_FILE_SIZE = 10000000;
+import {
+  createDropsForNFT,
+} from './nft-utils'
+
 const ACCEPTED_IMAGE_TYPES = ['image/jpeg', 'image/jpg', 'image/gif', 'image/png', 'image/webp'];
-export const DEBUG_DEL_NFT_ATTEMPT = async () => {
-  await del(NFT_ATTEMPT_KEY);
-};
 
 const schema = z.object({
   title: z.string().min(1, 'NFT name required'),
@@ -89,192 +82,9 @@ const CreateNftDropContext = createContext<CreateNftDropContextType>({
   },
 });
 
-const createDropsForNFT = async (dropId, returnTransactions, data, setAppModal) => {
-  const file = await data?.media?.arrayBuffer();
-
-  let { media = 'aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa' } = data;
-
-  const { numKeys, title, description } = data;
-
-  const wallet = await window.selector.wallet();
-
-  // generate CID locally HERE
-  if (file) {
-    const { root } = await pack({
-      input: file,
-      blockstore: new MemoryBlockStore(),
-      wrapWithDirectory: false,
-    });
-    media = root.toString();
-    console.log('CID', media);
-  }
-
-  let keys, requiredDeposit;
-  if (!data.seriesSecret) {
-    try {
-      const res = await createDrop({
-        wallet,
-        numKeys: 1,
-        metadata: JSON.stringify({
-          name: title,
-        }),
-        depositPerUseNEAR: 0.1,
-        fcData: {
-          methods: [
-            [
-              {
-                receiverId: 'nft-v2.keypom.testnet',
-                methodName: 'create_series',
-                args: JSON.stringify({
-                  mint_id: parseInt(dropId),
-                  metadata: {
-                    title,
-                    description,
-                    copies: numKeys,
-                    media,
-                  },
-                  // royalty?
-                }),
-                attachedDeposit: parseNearAmount('0.1')!,
-              },
-            ],
-          ],
-        },
-        useBalance: !returnTransactions,
-        returnTransactions,
-      });
-
-      keys = res.keys;
-      requiredDeposit = res.requiredDeposit;
-
-      if (!returnTransactions && !keys) {
-        throw new Error('Error creating drop');
-      }
-
-      // we're making the NFT now, so store the secret in case we have to re-attempt media upload
-      if (file) {
-        await update(NFT_ATTEMPT_KEY, (val) => ({ ...val, seriesSecret: keys.secretKeys[0] }));
-        data.seriesSecret = keys.secretKeys[0];
-      }
-    } catch (e) {
-      console.warn(e);
-      throw new Error('Error creating drop');
-    }
-  }
-
-  if (file) {
-    const { networkId } = getEnv();
-    const url = `${WORKER_BASE_URL}?network=${networkId as string}&secretKey=${
-      data.seriesSecret as string
-    }`;
-    let res;
-    try {
-      res = await fetch(url, {
-        method: 'POST',
-        body: file,
-      }).then(async (r) => await r.json());
-    } catch (error) {
-      console.warn('cfw error', error);
-      res = { error };
-    }
-
-    if (res.error) {
-      console.warn('cfw error', res.error);
-
-      const mediaErrorModal = () =>
-        setAppModal({
-          isOpen: true,
-          header: 'Whoops',
-          message: 'Your media failed to upload. Please reload this page to try again.',
-          options: [
-            {
-              label: 'Ok Reload',
-              func: () => {
-                window.location.reload();
-              },
-              buttonProps: {
-                variant: 'outline',
-              },
-            },
-            {
-              label: 'DEBUG - CANCEL',
-              func: async () => {
-                await del(NFT_ATTEMPT_KEY);
-                window.location.reload();
-              },
-              buttonProps: {
-                variant: 'outline',
-              },
-            },
-          ],
-        });
-
-      if (/Invalid drop/.test(res.error.toString())) {
-        return mediaErrorModal();
-      }
-
-      if (/drop not claimed/.test(res.error.toString())) {
-        return mediaErrorModal();
-      }
-
-      if (/media not uploaded/.test(res.error.toString())) {
-        // TODO get tx hash from error and store locally, resubmit tx hash to worker
-        // TODO worker verifies that it tried to claim and was successful
-        return mediaErrorModal();
-      }
-    }
-
-    await update(NFT_ATTEMPT_KEY, (val) => ({ ...val, seriesClaimed: true, fileUploaded: true }));
-
-    console.log('response from worker', res);
-  }
-
-  try {
-    const { publicKeys } = await generateKeys({
-      numKeys,
-      rootEntropy: `${localStorage.getItem(MASTER_KEY) as string}-${dropId as string}`,
-      autoMetaNonceStart: 0,
-    });
-
-    const { responses, requiredDeposit: requiredDeposit2 } = await createDrop({
-      wallet,
-      dropId,
-      numKeys,
-      publicKeys,
-      metadata: JSON.stringify({
-        dropName: title,
-      }),
-      fcData: {
-        methods: [
-          [
-            {
-              receiverId: 'nft-v2.keypom.testnet',
-              methodName: 'nft_mint',
-              args: '',
-              dropIdField: 'mint_id',
-              accountIdField: 'receiver_id',
-              attachedDeposit: parseNearAmount('0.1')!,
-            },
-          ],
-        ],
-      },
-      useBalance: !returnTransactions,
-      returnTransactions,
-    });
-
-    if (returnTransactions) {
-      return { requiredDeposit, requiredDeposit2 };
-    }
-
-    return { responses };
-  } catch (e) {
-    console.warn(e);
-    del(NFT_ATTEMPT_KEY);
-  }
-};
-
 export const CreateNftDropProvider = ({ children }: PropsWithChildren) => {
   const { data } = useSWRMutation('/api/drops/tokens', createLinks);
+
   const methods = useForm<Schema>({
     mode: 'onChange',
     defaultValues: {
@@ -349,7 +159,7 @@ export const CreateNftDropProvider = ({ children }: PropsWithChildren) => {
       title,
       description,
       numKeys,
-    });
+    }, null);
 
     const totalRequired = new BN(requiredDeposit).add(new BN(requiredDeposit2)).toString();
 
@@ -392,7 +202,7 @@ export const CreateNftDropProvider = ({ children }: PropsWithChildren) => {
     await addToBalance({
       wallet: await window.selector.wallet(),
       amountYocto: totalRequired.toString(),
-      successUrl: window.location.origin + '/drops',
+      successUrl: window.location.origin + '/drop/nft/new',
     });
   };
 
@@ -413,26 +223,6 @@ export const CreateNftDropProvider = ({ children }: PropsWithChildren) => {
       <FormProvider {...methods}>{children}</FormProvider>
     </CreateNftDropContext.Provider>
   );
-};
-
-export const handleFinishNFTDrop = async (setAppModal) => {
-  const data = (await get(NFT_ATTEMPT_KEY)) || {};
-  if (!data.dropId) {
-    return 
-  }
-
-  let res;
-  try {
-    res = await createDropsForNFT(data.dropId, false, data, setAppModal);
-  } catch (e) {
-    console.warn(e);
-  }
-
-  const { responses } = res;
-  console.log(responses);
-  if (responses.length > 0) {
-    del(NFT_ATTEMPT_KEY);
-  }
 };
 
 export const useCreateNftDrop = (): CreateNftDropContextType => {
